requirements.txt
fastapi
uvicorn
xgboost
scikit-learn
pandas
yfinance
requests
python-dotenv
groq
pydantic